Add SearchUser and RepoCommits components
