+++ src/App.css
@@ -1,3 +1,7 @@
 body {
   margin: 0px;
 }
+
+a {
+  text-decoration: none;
+}
