+++ src/Components/RepoCommits/repoCommits.css
@@ -1,0 +1,13 @@
+.repo-commits-item {
+  border-top: 1px solid rgb(184, 184, 184);
+  border-radius: 10px;
+  margin: 23px 15px;
+  /* display: flex;
+  flex-direction: column; */
+
+}
+
+.repo-commits-item p {
+  font-size: 14px;
+  margin: 8px 20px;
+}
