+++ src/Components/RepoItem/index.js
@@ -8,7 +8,7 @@ const RepoItem = ({ repo }) => {
     <div className="repo-item-container">
         <Link to={repo ? `/repo/${repo.owner.login}/${repo.name}` : "/"}> 
           <p><b>Description: </b>{repo.name}</p>
-          <p><b>Language: </b>{repo.language}</p> 
+          <p><b>Language: </b>{repo.language ? repo.language : ' - '}</p> 
         </Link>
     </div>
   )
